server/requirements.txt
Flask
python-dotenv
openai
langchain
langchain-openai
langchain-core
